// src/generated/sdk/models/portfolioCategoryProjectsByStatusOut.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * QMS Backend
 * OpenAPI spec version: 0.1.0
 */
import type { PortfolioCategoryCounts } from "./portfolioCategoryCounts";

export interface PortfolioCategoryProjectsByStatusOut {
  year: number;
  statuses?: string[];
  categories?: PortfolioCategoryCounts[];
}